requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pyzbar==0.1.9